pyproject.toml
[tool.poetry]
name = "insta_prepare"
version = "0.1.0"
description = "CLI tool for preparing images for Instagram publishing"
authors = ["Ilya Chichak <ilyachch@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
Pillow = "^8.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
